Show suggestion box of search facets when the search bar is in focus.
When the user focuses on the activity page search box a suggestion box is
shown underneath which contains possible query terms the user can enter.
When the input field loses focus the sugestion box is hidden.
The user can use the up/down arrows and the enter key to select suggestions.
The user can use the mouse to select suggestions from the dropdown box.
The user can use a combination of the up and down arrows and the mouse to
highlight a facet.

https://trello.com/c/I6s1UK1R/434-show-suggestion-box-under-search-box-on-focus